+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Fix false positive in `ReturnTypeDeclarationSniff` with nullable types.
 - Relax check for missing return type when `{aType}|null` doc bloc is present.
 - Add `is` to the list of allowed short names.
+- Added `FunctionBodyStartSniff` to deal with blank lines on top of functions body.
 
 ## 0.10.0
 - Renamed sniffs namespace (**breaking change**).